+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## x.y.z (unreleased)
 
 ### Added
+- Handling invalid xPath expression warnings
+  ([#361](https://github.com/jjriv/emogrifier/pull/361))
 
 
 ### Changed